requirements.txt
requests==2.31.0
python-dotenv==1.0.0
msal==1.24.1
pandas==2.1.3
openpyxl==3.1.2
python-dateutil==2.8.2
tqdm==4.66.1
webbrowser==3.11.0